Split server-listing-parents test into a folder of focused files
Break the 1,316-line server-listing-parents test into a
test/lib/server-listing-parents/ folder — children, edges, admin API,
add-on reachability, and deactivation & deletion — each well under the
400-line target, and drop its noExcessiveLinesPerFile exclusion from
biome.json.

Splitting surfaced duplication jscpd couldn't see inside the monolith.
Fixed it back to the 0% threshold with shared setup helpers
(parentAndChild, linkedParentChild, groupScopedAddOn, rescuingPageSetup,
soloChildAddOn) plus a curried edge-outcome assertion, and moved the
shared optInAddOnForListings helper into #test-utils (it was duplicated
in the bulk-actions deactivate test).
