+++ biome.json
@@ -155,7 +155,6 @@
         "src/features/api/payment-processing.ts",
         "src/shared/db/migrations/schema.ts",
         "src/ui/templates/public/reservations.tsx",
-        "test/lib/server-listing-parents.test.ts",
         "test/lib/server-payments.test.ts",
         "test/lib/square.test.ts",
         "test/lib/stripe.test.ts"
